stable version of Solomka

+++ ledger/src/shred/legacy.rs
@@ -423,7 +423,7 @@ mod test {
         {
             let mut shred = shred.clone();
             let index = shred.common_header.index - shred.common_header.fec_set_index - 1;
-            shred.set_index(index);
+            shred.set_index(index as u32);
             assert_matches!(
                 shred.sanitize(),
                 Err(Error::InvalidErasureShardIndex { .. })
